** Styles A methodology called BEM was used, which together with SMACSS and OOCSS is one of the most used in front-end development, in order to provide adequate maintenance of the style sheets in large projects.
Some mixins of my property were used that I often use in projects

It should be noted that for make the responsive it was used in this case for the simplicity and speed, flexbox without using a third-party framework

** Javascript
Regarding the javascript of the site, as it was minimal, it was decided to use Vanilla JS (Pure JS without libraries) since in this case no type of DOM management library or similar is justified.

** Wordpress Theme
In the wordpress theme all homepage has been added in the front-end.php file because it was necessary to add additional options using ACF.

ACF has been declared as a required plugin and cannot be deactivated, this is using the "TGM Plugin Activation" library which is according to the wordpress theming environment one of the best ways to add plugins to wordpress.
However, to avoid problems by not having the plugin installed, it will throw a " die() " until all the mandatory plugins are activated.

The wordpress theme exam is too short to extend it in a more complex hierarchy of folders and files, so it was decided to use only the functions.php file

To see the entire site as in the shared screenshots, it is then necessary to set a page as the front page so that the front-end.php file does its job.